show raw deal value numerically, weapon price increases

+++ Dragon's Trawel/src/trawel/personal/item/Item.java
@@ -121,7 +121,7 @@ public abstract class Item implements java.io.Serializable{
 	
 	public abstract String getName();
 	public abstract String getNameNoTier();
-	public abstract String storeString(float markup, int canShow);
+	public abstract String storeString(double markup, int canShow);
 	
 	protected void updateStats() {
